Extract terminal output behavior
