+++ src/app/components/workspace/terminal/terminal-viewport.tsx
@@ -19,6 +19,7 @@ import {
   terminalWrapperStyle,
   writeSystemMessage,
 } from './terminalViewportUtils'
+import { isXtermNearBottom, useTerminalOutputBehavior } from './useTerminalOutputBehavior'
 import {
   getTerminalPersistedSessionPath,
   useTerminalSessionLifecycle,
@@ -40,15 +41,6 @@ type TerminalViewportProps = {
   stickToBottomOnOutput?: boolean | undefined
   bottomAlignInitialContent?: boolean | undefined
   className?: string | undefined
-}
-
-const XTERM_STICKY_BOTTOM_THRESHOLD_ROWS = 2
-
-function isXtermNearBottom(terminal: XTerm) {
-  return (
-    terminal.buffer.active.baseY - terminal.buffer.active.viewportY <=
-    XTERM_STICKY_BOTTOM_THRESHOLD_ROWS
-  )
 }
 
 export function TerminalViewport({
@@ -73,8 +65,6 @@ export function TerminalViewport({
   const fitAddonRef = useRef<FitAddon | null>(null)
   const terminalResizeFrameRef = useRef<number | null>(null)
   const terminalResizeTimerRefs = useRef<number[]>([])
-  const pendingScrollFrameRef = useRef<number | null>(null)
-  const pendingBottomAlignFrameRef = useRef<number | null>(null)
   const terminalInitialFitTimerRef = useRef<number | null>(null)
   const sessionIdRef = useRef<string | null>(null)
   const attachFailedRef = useRef(false)
@@ -122,80 +112,11 @@ export function TerminalViewport({
     isFocused: isTerminalFocused,
   })
 
-  const scrollTerminalToBottom = useCallback(() => {
-    terminalInstanceRef.current?.scrollToBottom()
-  }, [])
-
-  const scheduleTerminalScrollToBottom = useCallback(() => {
-    if (pendingScrollFrameRef.current !== null) {
-      window.cancelAnimationFrame(pendingScrollFrameRef.current)
-    }
-
-    pendingScrollFrameRef.current = window.requestAnimationFrame(() => {
-      scrollTerminalToBottom()
-      pendingScrollFrameRef.current = window.requestAnimationFrame(() => {
-        scrollTerminalToBottom()
-        pendingScrollFrameRef.current = null
-      })
-    })
-  }, [scrollTerminalToBottom])
-
-  const applyXtermBottomAlign = useCallback(() => {
-    const terminal = terminalInstanceRef.current
-    const screen = terminal?.element?.querySelector<HTMLElement>('.xterm-screen')
-    if (!(terminal && screen)) {
-      return
-    }
-
-    if (!bottomAlignInitialContent || terminal.buffer.active.baseY > 0) {
-      screen.style.transform = ''
-      return
-    }
-
-    const screenHeight = screen.getBoundingClientRect().height
-    const rowHeight = terminal.rows > 0 ? screenHeight / terminal.rows : 0
-    if (!(Number.isFinite(rowHeight) && rowHeight > 0)) {
-      screen.style.transform = ''
-      return
-    }
-
-    const cursorY = terminal.buffer.active.cursorY
-    const offsetRows = Math.max(0, terminal.rows - cursorY - 1)
-    screen.style.transform = offsetRows > 0 ? `translateY(${offsetRows * rowHeight}px)` : ''
-  }, [bottomAlignInitialContent])
-
-  const scheduleXtermBottomAlign = useCallback(() => {
-    if (pendingBottomAlignFrameRef.current !== null) {
-      window.cancelAnimationFrame(pendingBottomAlignFrameRef.current)
-    }
-
-    pendingBottomAlignFrameRef.current = window.requestAnimationFrame(() => {
-      pendingBottomAlignFrameRef.current = window.requestAnimationFrame(() => {
-        pendingBottomAlignFrameRef.current = null
-        applyXtermBottomAlign()
-      })
-    })
-  }, [applyXtermBottomAlign])
-
-  const writeToTerminal = useCallback(
-    (data: string | Uint8Array) => {
-      const terminal = terminalInstanceRef.current
-      const shouldStickToBottom =
-        stickToBottomOnOutput && (!terminal || isXtermNearBottom(terminal))
-
-      terminal?.write(data, () => {
-        scheduleXtermBottomAlign()
-        if (shouldStickToBottom) {
-          terminal.scrollToBottom()
-        }
-      })
-
-      if (shouldStickToBottom) {
-        scheduleTerminalScrollToBottom()
-      }
-    },
-    [scheduleTerminalScrollToBottom, scheduleXtermBottomAlign, stickToBottomOnOutput],
-  )
+  const { scheduleXtermBottomAlign, writeToTerminal } = useTerminalOutputBehavior({
+    bottomAlignInitialContent,
+    stickToBottomOnOutput,
+    terminalInstanceRef,
+  })
 
   useEffect(
     () => () => {
@@ -206,12 +127,6 @@ export function TerminalViewport({
         window.clearTimeout(timer)
       }
       terminalResizeTimerRefs.current = []
-      if (pendingScrollFrameRef.current !== null) {
-        window.cancelAnimationFrame(pendingScrollFrameRef.current)
-      }
-      if (pendingBottomAlignFrameRef.current !== null) {
-        window.cancelAnimationFrame(pendingBottomAlignFrameRef.current)
-      }
       if (terminalInitialFitTimerRef.current !== null) {
         window.clearTimeout(terminalInitialFitTimerRef.current)
       }
